move snapshot to xds context
